Separate logic for different behaviour into compoenents
